fix: move the method isSolarDesignEnabled() to the abstract class ThemeServiceAbstract
