+++ models/classes/theme/ThemeServiceAbstract.php
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2017-2022 (original work) Open Assessment Technologies SA (under the project TAO-PRODUCT);
+ * Copyright (c) 2017-2024 (original work) Open Assessment Technologies SA (under the project TAO-PRODUCT);
  */
 
 namespace oat\tao\model\theme;
@@ -25,6 +25,7 @@ use oat\oatbox\service\ConfigurableService;
 use oat\tao\model\DynamicConfig\DynamicConfigProviderInterface;
 use oat\tao\model\featureFlag\FeatureFlagChecker;
 use oat\tao\model\featureFlag\FeatureFlagCheckerInterface;
+use Psr\Container\ContainerInterface;
 
 abstract class ThemeServiceAbstract extends ConfigurableService implements ThemeServiceInterface
 {
@@ -111,6 +112,13 @@ abstract class ThemeServiceAbstract extends ConfigurableService implements Theme
         }
 
         throw new common_exception_InconsistentData('The requested theme does not exist. (' . $themeId . ')');
+    }
+
+    public function isSolarDesignEnabled(): bool
+    {
+        return $this->getFeatureFlagChecker()->isEnabled(
+            FeatureFlagCheckerInterface::FEATURE_FLAG_SOLAR_DESIGN_ENABLED
+        );
     }
 
     /**
@@ -216,9 +224,17 @@ abstract class ThemeServiceAbstract extends ConfigurableService implements Theme
 
     protected function isTaoAsToolEnabled(): bool
     {
-        return $this->getServiceManager()
-            ->getContainer()
-            ->get(FeatureFlagChecker::class)
+        return $this->getFeatureFlagChecker()
             ->isEnabled(FeatureFlagCheckerInterface::FEATURE_FLAG_TAO_AS_A_TOOL);
+    }
+
+    protected function getFeatureFlagChecker(): FeatureFlagCheckerInterface
+    {
+        return $this->getContainer()->get(FeatureFlagChecker::class);
+    }
+
+    protected function getContainer(): ContainerInterface
+    {
+        return $this->getServiceManager()->getContainer();
     }
 }
